Add a property to disable the inter-process server
