+++ tests/src/test/java/net/devh/boot/grpc/test/setup/CustomCiphersAndProtocolsSetupTest.java
@@ -15,7 +15,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.test;
+package net.devh.boot.grpc.test.setup;
 
 import static org.junit.jupiter.api.Assertions.*;
