+++ tests/src/test/java/net/devh/boot/grpc/test/setup/InProcessSetupTest.java
@@ -15,7 +15,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.test;
+package net.devh.boot.grpc.test.setup;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
@@ -23,31 +23,22 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import lombok.extern.slf4j.Slf4j;
 import net.devh.boot.grpc.test.config.BaseAutoConfiguration;
+import net.devh.boot.grpc.test.config.InProcessConfiguration;
 import net.devh.boot.grpc.test.config.ServiceConfiguration;
 
 /**
- * A test checking that the server and client can start and connect to each other with minimal config.
+ * A test checking that the server and client can start and connect to each other in process.
  *
  * @author Daniel Theuke (daniel.theuke@heuboe.de)
  */
 @Slf4j
-@SpringBootTest(properties = {
-        "grpc.server.security.enabled=true",
-        "grpc.server.security.certificateChainPath=src/test/resources/certificates/server.crt",
-        "grpc.server.security.privateKeyPath=src/test/resources/certificates/server.key",
-        "grpc.server.security.trustCertCollectionPath=src/test/resources/certificates/client2.crt", // Wrong certs
-        "grpc.server.security.clientAuth=REQUIRE",
-        "grpc.client.test.security.authorityOverride=localhost",
-        "grpc.client.test.security.trustCertCollectionPath=src/test/resources/certificates/trusted-servers-collection",
-        "grpc.client.test.security.clientAuthEnabled=true",
-        "grpc.client.test.security.certificateChainPath=src/test/resources/certificates/client1.crt",
-        "grpc.client.test.security.privateKeyPath=src/test/resources/certificates/client1.key"})
-@SpringJUnitConfig(classes = {ServiceConfiguration.class, BaseAutoConfiguration.class})
+@SpringBootTest
+@SpringJUnitConfig(classes = {InProcessConfiguration.class, ServiceConfiguration.class, BaseAutoConfiguration.class})
 @DirtiesContext
-public class BrokenServerSelfSignedMutualSetupTest extends AbstractBrokenServerClientTest {
+public class InProcessSetupTest extends AbstractSimpleServerClientTest {
 
-    public BrokenServerSelfSignedMutualSetupTest() {
-        log.info("--- BrokenServerSelfSignedMutualSetupTest ---");
+    public InProcessSetupTest() {
+        log.info("--- InProcessSetupTest ---");
     }
 
 }
