+++ tests/src/test/java/net/devh/boot/grpc/test/setup/NameResolverConnectionTest.java
@@ -15,7 +15,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.test;
+package net.devh.boot.grpc.test.setup;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -26,18 +26,16 @@ import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import com.google.protobuf.Empty;
 
-import io.grpc.Status.Code;
 import net.devh.boot.grpc.client.inject.GrpcClient;
 import net.devh.boot.grpc.test.config.BaseAutoConfiguration;
 import net.devh.boot.grpc.test.config.ServiceConfiguration;
 import net.devh.boot.grpc.test.proto.TestServiceGrpc.TestServiceBlockingStub;
-import net.devh.boot.grpc.test.util.GrpcAssertions;
+import net.devh.boot.grpc.test.util.EnableOnIPv6;
 
 @SpringBootTest(properties = {
-        "grpc.server.address=127.0.0.1",
         "grpc.client.default.negotiationType=PLAINTEXT",
         "grpc.client.dns.negotiationType=PLAINTEXT",
-        "grpc.client.dns.address=dns:/localhost:9090/",
+        "grpc.client.dns.address=dns:///localhost:9090",
         "grpc.client.localhost.negotiationType=PLAINTEXT",
         "grpc.client.localhost.address=static://localhost:9090",
         "grpc.client.ipv4.negotiationType=PLAINTEXT",
@@ -47,20 +45,34 @@ import net.devh.boot.grpc.test.util.GrpcAssertions;
 })
 @SpringJUnitConfig(classes = {ServiceConfiguration.class, BaseAutoConfiguration.class})
 @DirtiesContext
-public class NameResolverIPv4ConnectionTest {
+public class NameResolverConnectionTest {
 
     private static final Empty EMPTY = Empty.getDefaultInstance();
 
+    @GrpcClient("default")
+    private TestServiceBlockingStub defaultStub;
     @GrpcClient("dns")
     private TestServiceBlockingStub dnsStub;
+    @GrpcClient("localhost")
+    private TestServiceBlockingStub localhostStub;
     @GrpcClient("ipv4")
     private TestServiceBlockingStub ipv4Stub;
     @GrpcClient("ipv6")
     private TestServiceBlockingStub ipv6Stub;
 
     @Test
+    public void testDefaultConnection() {
+        assertEquals("1.2.3", this.defaultStub.normal(EMPTY).getVersion());
+    }
+
+    @Test
     public void testDNSConnection() {
         assertEquals("1.2.3", this.dnsStub.normal(EMPTY).getVersion());
+    }
+
+    @Test
+    public void testLocalhostConnection() {
+        assertEquals("1.2.3", this.localhostStub.normal(EMPTY).getVersion());
     }
 
     @Test
@@ -69,8 +81,9 @@ public class NameResolverIPv4ConnectionTest {
     }
 
     @Test
+    @EnableOnIPv6
     public void testIpv6Connection() {
-        GrpcAssertions.assertThrowsStatus(Code.UNAVAILABLE, () -> this.ipv6Stub.normal(EMPTY));
+        assertEquals("1.2.3", this.ipv6Stub.normal(EMPTY).getVersion());
     }
 
 }
