+++ tests/src/test/java/net/devh/boot/grpc/test/setup/NameResolverIPv4ConnectionTest.java
@@ -15,9 +15,8 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.test;
+package net.devh.boot.grpc.test.setup;
 
-import static net.devh.boot.grpc.test.util.GrpcAssertions.assertThrowsStatus;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
@@ -32,12 +31,15 @@ import net.devh.boot.grpc.client.inject.GrpcClient;
 import net.devh.boot.grpc.test.config.BaseAutoConfiguration;
 import net.devh.boot.grpc.test.config.ServiceConfiguration;
 import net.devh.boot.grpc.test.proto.TestServiceGrpc.TestServiceBlockingStub;
-import net.devh.boot.grpc.test.util.EnableOnIPv6;
+import net.devh.boot.grpc.test.util.GrpcAssertions;
 
 @SpringBootTest(properties = {
-        "grpc.server.address=::1",
+        "grpc.server.address=127.0.0.1",
+        "grpc.client.default.negotiationType=PLAINTEXT",
         "grpc.client.dns.negotiationType=PLAINTEXT",
         "grpc.client.dns.address=dns:/localhost:9090/",
+        "grpc.client.localhost.negotiationType=PLAINTEXT",
+        "grpc.client.localhost.address=static://localhost:9090",
         "grpc.client.ipv4.negotiationType=PLAINTEXT",
         "grpc.client.ipv4.address=static://127.0.0.1:9090",
         "grpc.client.ipv6.negotiationType=PLAINTEXT",
@@ -45,8 +47,7 @@ import net.devh.boot.grpc.test.util.EnableOnIPv6;
 })
 @SpringJUnitConfig(classes = {ServiceConfiguration.class, BaseAutoConfiguration.class})
 @DirtiesContext
-@EnableOnIPv6
-public class NameResolverIPv6ConnectionTest {
+public class NameResolverIPv4ConnectionTest {
 
     private static final Empty EMPTY = Empty.getDefaultInstance();
 
@@ -64,12 +65,12 @@ public class NameResolverIPv6ConnectionTest {
 
     @Test
     public void testIpv4Connection() {
-        assertThrowsStatus(Code.UNAVAILABLE, () -> this.ipv4Stub.normal(EMPTY));
+        assertEquals("1.2.3", this.ipv4Stub.normal(EMPTY).getVersion());
     }
 
     @Test
     public void testIpv6Connection() {
-        assertEquals("1.2.3", this.ipv6Stub.normal(EMPTY).getVersion());
+        GrpcAssertions.assertThrowsStatus(Code.UNAVAILABLE, () -> this.ipv6Stub.normal(EMPTY));
     }
 
 }
