+++ tests/src/test/java/net/devh/boot/grpc/test/setup/PlaintextSetupTest.java
@@ -15,7 +15,7 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.devh.boot.grpc.test;
+package net.devh.boot.grpc.test.setup;
 
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
@@ -31,19 +31,13 @@ import net.devh.boot.grpc.test.config.ServiceConfiguration;
  * @author Daniel Theuke (daniel.theuke@heuboe.de)
  */
 @Slf4j
-@SpringBootTest(properties = {
-        "grpc.server.security.enabled=true",
-        "grpc.server.security.certificateChainPath=src/test/resources/certificates/server.crt",
-        "grpc.server.security.privateKeyPath=src/test/resources/certificates/server.key",
-        "grpc.client.test.security.authorityOverride=localhost",
-        "grpc.client.test.security.trustCertCollectionPath=src/test/resources/certificates/trusted-servers-collection"
-})
+@SpringBootTest(properties = "grpc.client.test.negotiationType=PLAINTEXT")
 @SpringJUnitConfig(classes = {ServiceConfiguration.class, BaseAutoConfiguration.class})
 @DirtiesContext
-public class SelfSignedServerSetupTest extends AbstractSimpleServerClientTest {
+public class PlaintextSetupTest extends AbstractSimpleServerClientTest {
 
-    public SelfSignedServerSetupTest() {
-        log.info("--- SelfSignedServerSetupTest ---");
+    public PlaintextSetupTest() {
+        log.info("--- PlaintextSetupTest ---");
     }
 
 }
